Added Add employee view

+++ src/main/java/org/up/roque/project/employee/ui/EmployeeTableModel.java
@@ -3,6 +3,7 @@ package org.up.roque.project.employee.ui;
 import org.up.roque.project.ProcessingException;
 import org.up.roque.project.employee.Employee;
 import org.up.roque.project.employee.EmployeeService;
+import org.up.roque.ui.DialogUtils;
 import org.up.roque.ui.MainFrame;
 import org.up.roque.ui.TableColumn;
 
@@ -13,15 +14,15 @@ import java.util.List;
 
 public class EmployeeTableModel extends AbstractTableModel {
   private final List<TableColumn<?, Employee>> columns = new ArrayList<>();
-  private final JFrame frame;
+  private final JFrame parentFrame;
   private final List<Employee> employees;
   private final EmployeeService service;
 
-  public EmployeeTableModel(JFrame frame, EmployeeService service) {
-    this.frame = frame;
+  public EmployeeTableModel(MainFrame frame, EmployeeService service) {
+    this.parentFrame = frame.getJFrame();
     this.service = service;
     this.employees = new ArrayList<>(service.findAll());
-    columns.add(new TableColumn<>("Id", Integer.class, Employee::getId));
+    columns.add(new TableColumn<>("Employee Id", Integer.class, Employee::getId));
     columns.add(new TableColumn<>("Name", String.class, Employee::getName));
     columns.add(new TableColumn<>("Cost Per Hour", Integer.class, Employee::getCostPerHour));
   }
@@ -54,14 +55,23 @@ public class EmployeeTableModel extends AbstractTableModel {
   }
 
   public void delete(int row) {
-    Employee employee = employees.get(row);
-    try {
-      service.delete(employee);
-    } catch (ProcessingException e) {
-      JOptionPane.showMessageDialog(frame, e.getMessage(), "Processing Error", JOptionPane.ERROR_MESSAGE);
+    if (row < 0) {
+      DialogUtils.error(parentFrame, "No row is selected");
       return;
     }
-    employees.remove(row);
-    fireTableDataChanged();
+    Employee employee = employees.get(row);
+    int response = DialogUtils.confirmation(parentFrame,
+        "Are you sure you want to delete employee '%s'?".formatted(employee.getName()), "Delete?");
+    if (response == JOptionPane.YES_OPTION) {
+      try {
+        service.delete(employee);
+      } catch (ProcessingException e) {
+        DialogUtils.error(parentFrame, e.getMessage());
+        return;
+      }
+      employees.remove(row);
+      fireTableDataChanged();
+    }
   }
+
 }
